Possible fix for clang warnings
git-svn-id: http://root.cern.ch/svn/root/trunk@39047 27541ba8-7e3a-0410-8455-c3a389f83636

+++ net/xrootd/src/xrootd/src/XrdFfs/XrdFfsQueue.cc
@@ -117,7 +117,7 @@ unsigned int XrdFfsQueue_count_tasks()
         else
 // this is wrong: 
 //          que_len = (unsigned int)2147483647 - (XrdFfsQueueTaskque_head->id - XrdFfsQueueTaskque_tail->id) + 1;
-            que_len = (unsigned int)4294967295 - (XrdFfsQueueTaskque_head->id - XrdFfsQueueTaskque_tail->id) + 1;
+            que_len = 4294967295u - (XrdFfsQueueTaskque_head->id - XrdFfsQueueTaskque_tail->id) + 1;
     }
     pthread_mutex_unlock(&XrdFfsQueueTaskque_mutex);
     return que_len;
